Project-Nexus/backend/requirements/base.txt
asgiref==3.11.0
attrs==25.4.0
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pathspec==0.12.1
Pillow==10.1.0
psycopg2-binary==2.9.7
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
setuptools==80.9.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
yamllint==1.37.1
